requirements.txt
pytube~=15.0.0
tqdm~=4.65.0
moviepy~=1.0.3
whisper
openai==0.27.8
tiktoken==0.3.1
python-dotenv==1.0.0